Move main component into addon and export facade in app
